feat : WinningLotto 클래스 삭제, LottoMachine 클래스의 역할을 분담할 WinningsStatistics 클래스 추가
